% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_detalhes.R
\name{pega_detalhes_processo}
\alias{pega_detalhes_processo}
\title{pega detalhes do empenho}
\usage{
pega_detalhes_processo(chave_primaria)
}
\arguments{
\item{chave_primaria}{codigo do empenho adquirido atraves da funcao pega_empenhos()}
}
\value{
dataset com os detalhes do empenho
}
\description{
pega detalhes do empenho
}
