% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_empenhos.R
\name{pega_empenhos}
\alias{pega_empenhos}
\title{pega empenhos}
\usage{
pega_empenhos(
  url =
    "https://formosodoaraguaia.megasofttransparencia.com.br/api/receitas-e-despesas/empenho/paginado"
)
}
\arguments{
\item{url}{url da api do portal da transparencia a ser raspado}
}
\value{
dataset com as informacoes a serem raspadas
}
\description{
pega empenhos
}
